fixed missing $ in , removed logging because of some permission thing

+++ tp2/run.sh
@@ -22,26 +22,28 @@ if [ "$OldInstances" != "" ]; then
         aws ec2 modify-instance-attribute --instance-id $instance --groups $DefaultSecurityGroup
     done
     aws ec2 terminate-instances --instance-ids $OldInstances
-    sleep 10
 fi
 
-OldGroups=$(aws ec2 describe-security-groups --query "SecurityGroups[].GroupId" --output text)
-for group in $OldGroups
-do
-    if [ "$group" != "$DefaultSecurityGroup" ]; then
-        aws ec2 delete-security-group --group-id $group
-    fi
-    sleep 10
-done
+SecurityGroup=$(aws ec2 describe-security-groups --query "SecurityGroups[].GroupId" --filter "Name=group-name,Values=tp2-group" --output text)
 
-SecurityGroup=$(aws ec2 create-security-group --description "tp2-group" --group-name tp2-group --output text)
-# enable inbound ssh to debug and http for us to view the webapp
-aws ec2 authorize-security-group-ingress --group-id $SecurityGroup --protocol tcp --port 22   --cidr 0.0.0.0/0
-aws ec2 authorize-security-group-ingress --group-id $SecurityGroup --protocol tcp --port 80   --cidr 0.0.0.0/0
-aws ec2 authorize-security-group-ingress --group-id $SecurityGroup --protocol tcp --port 443  --cidr 0.0.0.0/0
-# for downloads, enable http/https outbound
-aws ec2 authorize-security-group-egress  --group-id $SecurityGroup --protocol tcp --port 80   --cidr 0.0.0.0/0
-aws ec2 authorize-security-group-egress  --group-id $SecurityGroup --protocol tcp --port 443  --cidr 0.0.0.0/0
+if [ "$SecurityGroup" == "" ]; then
+    OldGroups=$(aws ec2 describe-security-groups --query "SecurityGroups[].GroupId" --output text)
+    for group in $OldGroups
+    do
+        if [ "$group" != "$DefaultSecurityGroup" ]; then
+            aws ec2 delete-security-group --group-id $group
+        fi
+        sleep 10
+    done
+    SecurityGroup=$(aws ec2 create-security-group --description "tp2-group" --group-name tp2-group --output text)
+    # enable inbound ssh to debug and http for us to view the webapp
+    aws ec2 authorize-security-group-ingress --group-id $SecurityGroup --protocol tcp --port 22   --cidr 0.0.0.0/0
+    aws ec2 authorize-security-group-ingress --group-id $SecurityGroup --protocol tcp --port 80   --cidr 0.0.0.0/0
+    aws ec2 authorize-security-group-ingress --group-id $SecurityGroup --protocol tcp --port 443  --cidr 0.0.0.0/0
+    # for downloads, enable http/https outbound
+    aws ec2 authorize-security-group-egress  --group-id $SecurityGroup --protocol tcp --port 80   --cidr 0.0.0.0/0
+    aws ec2 authorize-security-group-egress  --group-id $SecurityGroup --protocol tcp --port 443  --cidr 0.0.0.0/0
+fi
 
 M4Large="$(aws ec2 run-instances --image-id $ECSImageId --count 1 --instance-type m4.large --security-group-ids $SecurityGroup --key-name vockey --user-data file://setupInstance.sh --placement AvailabilityZone=$Zone --query "Instances[].[InstanceId]" --output text)"
 echo $M4Large
